Package refactor of login code and beginning of router modularization
