+++ api/src/main/java/com/codeforcommunity/api/IAuthProcessor.java
@@ -1,5 +1,10 @@
 package com.codeforcommunity.api;
 
+import com.codeforcommunity.dto.auth.IsUserRequest;
+import com.codeforcommunity.dto.auth.NewSessionRequest;
+import com.codeforcommunity.dto.auth.NewUserRequest;
+import com.codeforcommunity.dto.auth.RefreshSessionRequest;
+import com.codeforcommunity.dto.auth.RefreshSessionResponse;
 import com.codeforcommunity.exceptions.AuthException;
 import com.codeforcommunity.dto.*;
