--- api/src/main/java/com/codeforcommunity/dto/MemberReturn.java
@@ -1,29 +1,0 @@
-package com.codeforcommunity.dto;
-
-public class MemberReturn {
-  private String firstName;
-  private String lastName;
-
-  public MemberReturn(String firstName, String lastName) {
-    this.firstName = firstName;
-    this.lastName = lastName;
-  }
-
-  private MemberReturn() {}
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public void setFirstName(String firstName) {
-    this.firstName = firstName;
-  }
-
-  public String getLastName() {
-    return lastName;
-  }
-
-  public void setLastName(String lastName) {
-    this.lastName = lastName;
-  }
-}
