+++ api/src/main/java/com/codeforcommunity/dto/auth/NewSessionRequest.java
@@ -1,4 +1,6 @@
-package com.codeforcommunity.dto;
+package com.codeforcommunity.dto.auth;
+
+import com.codeforcommunity.dto.IDTO;
 
 public class NewSessionRequest implements IDTO {
     private String username;
