+++ api/src/main/java/com/codeforcommunity/dto/auth/RefreshSessionRequest.java
@@ -1,4 +1,4 @@
-package com.codeforcommunity.dto;
+package com.codeforcommunity.dto.auth;
 
 public class RefreshSessionRequest {
     private String refreshToken;
