+++ api/src/main/java/com/codeforcommunity/dto/notes/NoteRequest.java
@@ -1,4 +1,4 @@
-package com.codeforcommunity.dto;
+package com.codeforcommunity.dto.notes;
 
 public class NoteRequest {
