+++ api/src/main/java/com/codeforcommunity/rest/IRouter.java
@@ -1,0 +1,9 @@
+package com.codeforcommunity.rest;
+
+import io.vertx.core.Vertx;
+import io.vertx.ext.web.Router;
+
+public interface IRouter {
+
+  Router initializeRouter(Vertx vertx);
+}
